+++ wfd-xml/impl/src/main/java/com/quadient/wfdxml/WfdXmlBuilder.java
@@ -47,9 +47,20 @@ public class WfdXmlBuilder {
     }
 
     public String build() {
+        return this.build(false);
+    }
+
+    public String build(boolean withDeltaStyles) {
         XmlExporter exporter = new XmlExporter();
         exporter.declaration("1.0", "UTF-8");
         exporter.beginElement("WorkFlow");
+
+        if (withDeltaStyles) {
+            exporter.beginElement("Property");
+            exporter.addElementWithStringData("Name", "DeltaStyles");
+            exporter.addElementWithIntData("Value", 1);
+            exporter.endElement();
+        }
 
         for (WorkFlowModuleImpl module : modules) {
             module.export(exporter);
@@ -60,6 +71,17 @@ public class WfdXmlBuilder {
         }
 
         exporter.endElement();
+        return exporter.buildString();
+    }
+
+    public String buildStyleLayoutDelta() {
+        XmlExporter exporter = new XmlExporter();
+
+        Optional<LayoutImpl> layoutModule = modules.stream().filter(LayoutImpl.class::isInstance).map(LayoutImpl.class::cast).findFirst();
+        if (layoutModule.isEmpty()) throw new IllegalStateException("No layout module found");
+
+        layoutModule.get().exportStyleLayoutDelta(exporter);
+
         return exporter.buildString();
     }