+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 ### Added
 
 - Added sample input files for DocBook example to the resources folder
+- deployStyles now creates wfd as well as jld style definition for interactive
 
 ### Changed